Add calculator logic and display functions
Added function 'calculator' which serves as a callback function when
adding event listeners to all divs of class 'button'.

'Calculator' serves as a logical roadmap that directs the application to
the appropriate functions to call depending on where the user is
currently at in their calculation.

Function setFirstOperand, like the name says, sets the first operand in
the expression depending on what number the user has entered on the
calculator.

Function setOperator sets the operator depending on what operator the
user picked. Also evaluates the previous operand-operator pair if user
entered one before choosing another operator (for example, if user
enters 2 + 2, and then presses another operator button like +, -, etc.
setOperator will evaulate 2 + 2, and 4 will be the new number that the
operator is operating on).

Function setSecondOperand sets the second operand in the expression
depending on what number the user has entered on the calculator.

Function updateDisplay changes the display whenever the user enters a
number or evaluates a operand-operator pair.

Function resetCalc resets the state of the calculator, wiping all
operator and operand data. This is called when clicking the clear (AC)
button or clicking a number after evaluating an operand-operator pair
with the = button.

Function resetDisplay is called alongside resetCalc and sets the display
to read 0.